requirements.txt
jupyter>=1.0.0
notebook>=7.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
nltk>=3.8.0
wordcloud>=1.9.0
xgboost>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.0
wordninja>=2.0.0

# PDF Extraction Libraries
pdfplumber>=0.10.0
PyPDF2>=3.0.0
tabula-py>=2.8.0
camelot-py>=0.11.0

# Date Parsing
python-dateutil>=2.8.0

# Excel Output
openpyxl>=3.1.0

# Optional: OCR for scanned PDFs (requires Tesseract installation)
# pytesseract>=0.3.10
# pdf2image>=1.16.0

# TensorFlow for Neural Network
tensorflow>=2.13.0

